+++ userprofile/forms.py
@@ -1,8 +1,16 @@
 from django import forms
 
-from projectroles.models import PROJECT_SETTING_VAL_MAXLENGTH
-from projectroles.plugins import ProjectAppPluginPoint
-from projectroles.user_settings import get_user_setting, validate_user_setting
+from projectroles.app_settings import AppSettingAPI
+from projectroles.models import APP_SETTING_VAL_MAXLENGTH, SODAR_CONSTANTS
+from projectroles.plugins import get_active_plugins
+
+
+# SODAR Constants
+APP_SETTING_SCOPE_USER = SODAR_CONSTANTS['APP_SETTING_SCOPE_USER']
+
+
+# App settings API
+app_settings = AppSettingAPI()
 
 
 # User Settings Form -----------------------------------------------------------
@@ -18,59 +26,57 @@ class UserSettingsForm(forms.Form):
         super().__init__(*args, **kwargs)
 
         # Add settings fields
-        self.app_plugins = list(
-            sorted(
-                [
-                    p
-                    for p in ProjectAppPluginPoint.get_plugins()
-                    if p.user_settings
-                ],
-                key=lambda x: x.name,
-            )
-        )
+        self.app_plugins = get_active_plugins()
 
-        for p in self.app_plugins:
-            for s_key in sorted(p.user_settings):
-                s = p.user_settings[s_key]
-                s_field = 'settings.{}.{}'.format(p.name, s_key)
+        for plugin in self.app_plugins:
+            p_settings = app_settings.get_setting_defs(
+                plugin, APP_SETTING_SCOPE_USER
+            )
+
+            for s_key, s_val in p_settings.items():
+                s_field = 'settings.{}.{}'.format(plugin.name, s_key)
                 field_kwarg = {
                     'required': False,
-                    'label': s.get('label') or '{}.{}'.format(p.name, s_key),
-                    'help_text': s.get('description'),
+                    'label': s_val.get('label')
+                    or '{}.{}'.format(plugin.name, s_key),
+                    'help_text': s_val.get('description'),
                 }
-                widget_attrs = {'placeholder': s.get('placeholder') or None}
+                widget_attrs = {'placeholder': s_val.get('placeholder') or None}
 
-                if s['type'] == 'STRING':
+                if s_val['type'] == 'STRING':
                     self.fields[s_field] = forms.CharField(
-                        max_length=PROJECT_SETTING_VAL_MAXLENGTH,
+                        max_length=APP_SETTING_VAL_MAXLENGTH,
                         widget=forms.TextInput(attrs=widget_attrs),
                         **field_kwarg,
                     )
 
-                elif s['type'] == 'INTEGER':
+                elif s_val['type'] == 'INTEGER':
                     self.fields[s_field] = forms.IntegerField(
                         widget=forms.TextInput(attrs=widget_attrs),
                         **field_kwarg,
                     )
 
-                elif s['type'] == 'BOOLEAN':
+                elif s_val['type'] == 'BOOLEAN':
                     self.fields[s_field] = forms.BooleanField(**field_kwarg)
 
                 # Set initial value
-                self.initial[s_field] = get_user_setting(
-                    user=self.user, app_name=p.name, setting_name=s_key
+                self.initial[s_field] = app_settings.get_app_setting(
+                    app_name=plugin.name, setting_name=s_key, user=self.user
                 )
 
     def clean(self):
         """Function for custom form validation and cleanup"""
 
-        for p in self.app_plugins:
-            for s_key in sorted(p.user_settings):
-                s = p.user_settings[s_key]
-                s_field = 'settings.{}.{}'.format(p.name, s_key)
+        for plugin in self.app_plugins:
+            p_settings = app_settings.get_setting_defs(
+                plugin, APP_SETTING_SCOPE_USER
+            )
 
-                if not validate_user_setting(
-                    setting_type=s['type'],
+            for s_key, s_val in p_settings.items():
+                s_field = 'settings.{}.{}'.format(plugin.name, s_key)
+
+                if not app_settings.validate_setting(
+                    setting_type=s_val['type'],
                     setting_value=self.cleaned_data.get(s_field),
                 ):
                     self.add_error(s_field, 'Invalid value')