+++ userprofile/tests/test_views.py
@@ -5,8 +5,12 @@ from django.test import RequestFactory
 
 from test_plus.test import TestCase
 
-from projectroles.tests.test_models import EXAMPLE_APP_NAME, ProjectSettingMixin
-from projectroles.user_settings import get_user_setting
+from projectroles.app_settings import AppSettingAPI
+from projectroles.tests.test_models import EXAMPLE_APP_NAME, AppSettingMixin
+
+
+# App settings API
+app_settings = AppSettingAPI()
 
 
 class TestViewsBase(TestCase):
@@ -37,7 +41,7 @@ class TestUserDetailView(TestViewsBase):
         self.assertIsNotNone(response.context['user_settings'])
 
 
-class TestUserSettingsForm(ProjectSettingMixin, TestViewsBase):
+class TestUserSettingsForm(AppSettingMixin, TestViewsBase):
     """Tests for the user settings form."""
 
     # NOTE: This assumes an example app is available
@@ -48,7 +52,7 @@ class TestUserSettingsForm(ProjectSettingMixin, TestViewsBase):
         # Init test setting
         self.setting_str = self._make_setting(
             app_name=EXAMPLE_APP_NAME,
-            name='str_setting',
+            name='user_str_setting',
             setting_type='STRING',
             value='test',
             user=self.user,
@@ -57,7 +61,7 @@ class TestUserSettingsForm(ProjectSettingMixin, TestViewsBase):
         # Init integer setting
         self.setting_int = self._make_setting(
             app_name=EXAMPLE_APP_NAME,
-            name='int_setting',
+            name='user_int_setting',
             setting_type='INTEGER',
             value=170,
             user=self.user,
@@ -66,7 +70,7 @@ class TestUserSettingsForm(ProjectSettingMixin, TestViewsBase):
         # Init boolean setting
         self.setting_bool = self._make_setting(
             app_name=EXAMPLE_APP_NAME,
-            name='bool_setting',
+            name='user_bool_setting',
             setting_type='BOOLEAN',
             value=True,
             user=self.user,
@@ -74,45 +78,54 @@ class TestUserSettingsForm(ProjectSettingMixin, TestViewsBase):
 
     def testGet(self):
         with self.login(self.user):
-            response = self.client.get(reverse('userprofile:update-settings'))
+            response = self.client.get(reverse('userprofile:settings_update'))
         self.assertEqual(response.status_code, 200)
         self.assertIsNotNone(response.context['form'])
         self.assertIsNotNone(
             response.context['form'].fields.get(
-                'settings.%s.str_setting' % EXAMPLE_APP_NAME
+                'settings.%s.user_str_setting' % EXAMPLE_APP_NAME
             )
         )
         self.assertIsNotNone(
             response.context['form'].fields.get(
-                'settings.%s.int_setting' % EXAMPLE_APP_NAME
+                'settings.%s.user_int_setting' % EXAMPLE_APP_NAME
             )
         )
         self.assertIsNotNone(
             response.context['form'].fields.get(
-                'settings.%s.bool_setting' % EXAMPLE_APP_NAME
+                'settings.%s.user_bool_setting' % EXAMPLE_APP_NAME
             )
         )
 
     def testPost(self):
         self.assertEqual(
-            get_user_setting(self.user, EXAMPLE_APP_NAME, 'str_setting'), 'test'
+            app_settings.get_app_setting(
+                EXAMPLE_APP_NAME, 'user_str_setting', user=self.user
+            ),
+            'test',
         )
         self.assertEqual(
-            get_user_setting(self.user, EXAMPLE_APP_NAME, 'int_setting'), 170
+            app_settings.get_app_setting(
+                EXAMPLE_APP_NAME, 'user_int_setting', user=self.user
+            ),
+            170,
         )
         self.assertEqual(
-            get_user_setting(self.user, EXAMPLE_APP_NAME, 'bool_setting'), True
+            app_settings.get_app_setting(
+                EXAMPLE_APP_NAME, 'user_bool_setting', user=self.user
+            ),
+            True,
         )
 
         values = {
-            'settings.%s.str_setting' % EXAMPLE_APP_NAME: 'another-text',
-            'settings.%s.int_setting' % EXAMPLE_APP_NAME: '123',
-            'settings.%s.bool_setting' % EXAMPLE_APP_NAME: False,
+            'settings.%s.user_str_setting' % EXAMPLE_APP_NAME: 'another-text',
+            'settings.%s.user_int_setting' % EXAMPLE_APP_NAME: '123',
+            'settings.%s.user_bool_setting' % EXAMPLE_APP_NAME: False,
         }
 
         with self.login(self.user):
             response = self.client.post(
-                reverse('userprofile:update-settings'), values
+                reverse('userprofile:settings_update'), values
             )
 
         # Assert redirect
@@ -121,12 +134,20 @@ class TestUserSettingsForm(ProjectSettingMixin, TestViewsBase):
 
         # Assert settings state after update
         self.assertEqual(
-            get_user_setting(self.user, EXAMPLE_APP_NAME, 'str_setting'),
+            app_settings.get_app_setting(
+                EXAMPLE_APP_NAME, 'user_str_setting', user=self.user
+            ),
             'another-text',
         )
         self.assertEqual(
-            get_user_setting(self.user, EXAMPLE_APP_NAME, 'int_setting'), 123
+            app_settings.get_app_setting(
+                EXAMPLE_APP_NAME, 'user_int_setting', user=self.user
+            ),
+            123,
         )
         self.assertEqual(
-            get_user_setting(self.user, EXAMPLE_APP_NAME, 'bool_setting'), False
+            app_settings.get_app_setting(
+                EXAMPLE_APP_NAME, 'user_bool_setting', user=self.user
+            ),
+            False,
         )